Split updater to separate program
