+++ src/commonMain/kotlin/dev/zacsweers/ReadMeTemplate.kt
@@ -1,5 +1,7 @@
 package dev.zacsweers
 
+import dev.zacsweers.ReadmeUpdater.ActivityItem
+
 fun createReadMe(githubActivity: List<ActivityItem>, blogActivity: List<ActivityItem>): String {
   return """
     Currently working on [Slack](https://slack.com/). Read [my blog](https://zacsweers.dev/) or follow me on places `@ZacSweers`.
